+++ data_classes/CHANGELOG.md
@@ -1,16 +1,28 @@
-## 1.1.0 - 2019-09-05
+## [2.0.1] - 2019-09-20
+
+- Revise readme: Little typo fixes and document `build_runner` dependency.
+
+## [2.0.0] - 2019-09-20
+
+- Change `@DataClass()` annotation to `@GenerateDataClassFor()`.
+- `GeneratedClass.fromMutable()` is now a normal constructor instead of a
+  factory constructor.
+- Provide new example.
+- Revise readme.
+- New license.
+
+## [1.1.0] - 2019-09-05
 
 - Change `@Nullable()` annotation to `@nullable`.
 
-## 1.0.2 - 2019-09-04
+## [1.0.2] - 2019-09-04
 
-- Rename example's mutable class to `MutableUser`. Also adjust the readme
-  accordingly.
+- Rename example's mutable class to `MutableUser`.
 
-## 1.0.1 - 2019-09-04
+## [1.0.1] - 2019-09-04
 
 - Added example.
 
-## 1.0.0 - 2019-09-04
+## [1.0.0] - 2019-09-04
 
 - Initial release: Supporting `DataClass` and `Nullable` annotations.